+++ CHANGELOG.md
@@ -15,11 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - generic filter: `exclude_kicad_dnp` to exclude components marked as DNP in
   KiCad (See #887)
 - KiCad 9 embedded files: populate the cache if needed (See #890)
+- 3D models: option to download the latest 3D models (`kicad_3d_url_version`)
+  (See #899)
 - BoM: descriptive error when using `kicad` format and no options found.
 - DRC/ERC: Configurable logo (#889)
 - iBoM: `mark_when_checked` option (#881)
 
 ### Fixed
+- 3D models: download models from the repo tag corresponding to the currently
+  installed KiCad version, not the one from `master` (#899)
 - Blender export: problems with component values containing `/ ` (#885)
 - Panelize:
   - Problems when using angles (#898)