+++ docs/source/Changelog.rst
@@ -32,6 +32,8 @@ Added
 -  generic filter: ``exclude_kicad_dnp`` to exclude components marked as
    DNP in KiCad (See #887)
 -  KiCad 9 embedded files: populate the cache if needed (See #890)
+-  3D models: option to download the latest 3D models
+   (``kicad_3d_url_version``) (See #899)
 -  BoM: descriptive error when using ``kicad`` format and no options
    found.
 -  DRC/ERC: Configurable logo (#889)
@@ -40,6 +42,8 @@ Added
 Fixed
 ~~~~~
 
+-  3D models: download models from the repo tag corresponding to the
+   currently installed KiCad version, not the one from ``master`` (#899)
 -  Blender export: problems with component values containing ``/``
    (#885)
 -  Panelize: